requirements.txt
Django>=5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
drf-yasg>=1.21
psycopg2-binary>=2.9
